// src/components/ConsequenceTracer.module.css.d.ts
declare const styles: {
  consequenceTracer: string;
  header: string;
  title: string;
  subtitle: string;
  timeline: string;
  timelineGroup: string;
  step: string;
  "step-blue": string;
  "step-yellow": string;
  "step-red": string;
  "step-green": string;
  stepIcon: string;
  stepContent: string;
  stepTitle: string;
  stepDescription: string;
  connector: string;
};

export = styles;